api/requirements.txt
alembic==1.4.3
aniso8601==8.0.0
blinker==1.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
PyJWT==1.7.1
pymongo==3.11.0
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.19
Werkzeug==1.0.1
